+++ docs/changelog.md
@@ -1,6 +1,21 @@
 Change Log
 ==========
 
+Version 0.5.3 (2024-04-25)
+--------------------------
+
+### Dependencies
+
+- Updated Android Gradle Plugin to 8.3.2.
+- Updated `apksig` to 8.3.2.
+- Added Apache Commons Compress dependency to `ackpine-splits` module.
+
+### Bug fixes and improvements
+
+- Use `FileChannel` to read zipped APKs on Android Oreo+ if possible. This drastically improves performance when direct access through `java.io` APIs is not available and allows to process problematic ZIP files (such as XAPK files).
+- Don't crash if exception occurs while iterating APK sequence in sample apps, and display the exception message instead.
+- Share one thread pool across ViewModels in Java sample.
+
 Version 0.5.2 (2024-03-30)
 --------------------------